Add TaskListKind Ephemeral (#209)
Add support for Ephemeral TaskLists, which will automatically be deleted once no longer in use. This is behavior already supported by Sticky TaskLists, but non-sticky TaskList scenarios such as sessions and integration tests both require generating dynamic TaskLists. We need similar functionality for these TaskLists.

Requests to Start/SignalWithStart a workflow specify a TaskList instance (Name + Kind), so no changes are needed for those.

Add a TaskList to the response of DescribeTaskList so that the server can expose the TaskListKind of a given TaskList. The request currently includes a TaskList, but the TaskListKind is not considered part of the identity of a TaskList and so the Kind field there is ignored. As a result there isn't a way to determine the Kind of a TaskList beyond inferring it from the name or looking at the database.

Add a TaskListKind field to WorkflowExecutionInfo. This allows for DescribeWorkflow to return the TaskListKind the workflow was started with. Currently this is implicitly TaskListKind NORMAL.

Update sqlblobs WorkflowExecutionInfo so that we can store it for SQL persistence types.